refactor(foundation): move contract_drift value types (A6)
Phase 6 / Sub-phase A6: Extract the data-only types from contract_drift/types.rs
to foundation, now that ai_contract_diff types (MismatchSeverity, MismatchType)
are also in foundation.

Moved to mockforge-foundation::contract_drift_types:
- DriftBudget
- DriftBudgetConfig
- DriftMetrics
- BreakingChangeRule, BreakingChangeRuleType, BreakingChangeRuleConfig
- BreakingChangeRule::matches() method (pure logic, no cross-module deps)

Left in mockforge-core (has cross-module deps):
- DriftResult — uses FitnessTestResult (contract_drift::fitness) and
  ConsumerImpact (contract_drift::consumer_mapping) fields
- DriftResult::from_diff_result() — uses BreakingChangeDetector

Re-exports preserved via core's contract_drift/types.rs so external consumers
(`mockforge_core::contract_drift::types::DriftBudget`, etc.) still work.

Workspace builds clean.
